# swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_User.Exclude_keyofUser.password_hash-or-refresh_token_version__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"id": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"email_verified": {
						"type": "boolean"
					},
					"email": {
						"type": "string"
					},
					"photo": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"bio",
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.password_hash-or-refresh_token_version_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password_hash-or-refresh_token_version__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"SanitizedUser": {
				"$ref": "#/components/schemas/Omit_User.password_hash-or-refresh_token_version_"
			},
			"APIResponse_SanitizedUser_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/SanitizedUser"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginationResponse_SanitizedUser_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"list": {
						"items": {
							"$ref": "#/components/schemas/SanitizedUser"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"list"
				],
				"type": "object"
			},
			"APIResponse_PaginationResponse_SanitizedUser__": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PaginationResponse_SanitizedUser_"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserFetchList": {
				"properties": {
					"limit": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Expand_Optional_UserMutable.bio__": {
				"properties": {
					"bio": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"photo": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"role",
					"email",
					"photo"
				],
				"type": "object"
			},
			"UserCreate": {
				"$ref": "#/components/schemas/Expand_Optional_UserMutable.bio__"
			},
			"Partial_Omit_UserMutable.email__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"photo": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UserUpdate": {
				"$ref": "#/components/schemas/Partial_Omit_UserMutable.email__"
			},
			"APIResponse__url-string__": {
				"properties": {
					"data": {
						"properties": {
							"url": {
								"type": "string"
							}
						},
						"required": [
							"url"
						],
						"type": "object",
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			},
			"APIResponse_Message_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Message"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthLoginResponse": {
				"properties": {
					"refreshToken": {
						"type": "string"
					},
					"accessToken": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/SanitizedUser"
					}
				},
				"required": [
					"refreshToken",
					"accessToken",
					"user"
				],
				"type": "object"
			},
			"APIResponse_AuthLoginResponse_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AuthLoginResponse"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthLogin": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"AuthVerifyEmail": {
				"properties": {
					"email": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"otp"
				],
				"type": "object"
			},
			"AuthForgotPass": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object"
			},
			"AuthResetPass": {
				"properties": {
					"password": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"otp",
					"email"
				],
				"type": "object"
			},
			"AuthChangePass": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"AuthTemplateType": {
				"type": "string",
				"enum": [
					"verifyEmail",
					"forgotPassword",
					"createUser"
				],
				"nullable": false
			},
			"AuthVerification": {
				"properties": {
					"emailType": {
						"$ref": "#/components/schemas/AuthTemplateType"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"emailType",
					"email"
				],
				"type": "object"
			},
			"AuthTokenResponse": {
				"properties": {
					"refreshToken": {
						"type": "string"
					},
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken",
					"accessToken"
				],
				"type": "object"
			},
			"APIResponse_AuthTokenResponse_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AuthTokenResponse"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "A resource owner JWT"
			}
		}
	},
	"info": {
		"title": "rayon_express_psql_starter",
		"version": "1.2",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/{userId}": {
			"get": {
				"operationId": "Fetch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_SanitizedUser_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_SanitizedUser_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_SanitizedUser_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "FetchList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_PaginationResponse_SanitizedUser__"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_SanitizedUser_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/UserCreate"
									},
									{
										"properties": {
											"password": {
												"type": "string"
											}
										},
										"required": [
											"password"
										],
										"type": "object"
									}
								]
							}
						}
					}
				}
			}
		},
		"/files": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse__url-string__"
								}
							}
						}
					}
				},
				"tags": [
					"Files"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_Message_"
								}
							}
						}
					}
				},
				"tags": [
					"Files"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"url": {
										"type": "string"
									}
								},
								"required": [
									"url"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_AuthLoginResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthLogin"
							}
						}
					}
				}
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_AuthLoginResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"bio": {
										"type": "string"
									},
									"photo": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"name",
									"email",
									"password"
								]
							}
						}
					}
				}
			}
		},
		"/auth/verifyEmail": {
			"post": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_Message_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthVerifyEmail"
							}
						}
					}
				}
			}
		},
		"/auth/signoutAll": {
			"post": {
				"operationId": "Signout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_Message_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/forgotPassword": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_Message_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthForgotPass"
							}
						}
					}
				}
			}
		},
		"/auth/resetPassword": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_Message_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthResetPass"
							}
						}
					}
				}
			}
		},
		"/auth/changePassword": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_Message_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthChangePass"
							}
						}
					}
				}
			}
		},
		"/auth/resendVerification": {
			"post": {
				"operationId": "ResendVerification",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_Message_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthVerification"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_AuthTokenResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}